Some changes and updates.

+++ HomeShark.Core/Models/Station.cs
@@ -21,9 +21,9 @@ namespace HomeShark.Core.Models
 
         public StationType StationType { get; set; }
 
-        public double Latitude { get; set; }
+        public long Latitude { get; set; }
 
-        public double Longitude { get; set; }
+        public long Longitude { get; set; }
 
         public string GeoLocation => $"{Latitude},{Longitude}";
